ml/inference/requirements.txt
torch
facenet_pytorch
pillow
numpy<2
scikit-learn
tqdm
fastapi
uvicorn[standard]